Change eslint rules
I think preferring the default export doesn’t make a whole lot of sense for us at the moment, because our code is changing so much that swapping between default and non-default exports is going to happen pretty often, which will lead to extra code change.

No-use-before-define should not apply to functions, as those are specifically hoisted in javascript.