Add dummy event listener code
I’m not sure yet what will need to happen for each of these events, so I’ve added a dummy structure for the moment.

There’s also a function for disabling the listeners when they’re no longer needed – for the moment I’ve made that when a "gameEnd" event is received.